DRY and visibility icon + reminder order bug fixes for Create/Edit Event modals
