+++ plugins/calendar-resources/src/components/CreateEvent.svelte
@@ -16,31 +16,12 @@
   import { Calendar, RecurringRule, Visibility, generateEventId } from '@hcengineering/calendar'
   import { Person, PersonAccount } from '@hcengineering/contact'
   import { Class, Doc, Ref, getCurrentAccount } from '@hcengineering/core'
-  import presentation, { createQuery, getClient } from '@hcengineering/presentation'
-  import { StyledTextBox } from '@hcengineering/text-editor'
-  import {
-    Button,
-    EditBox,
-    FocusHandler,
-    Icon,
-    IconClose,
-    IconMoreH,
-    createFocusManager,
-    getUserTimezone,
-    showPopup,
-    Scroller
-  } from '@hcengineering/ui'
+  import { createQuery, getClient } from '@hcengineering/presentation'
+  import { getUserTimezone } from '@hcengineering/ui'
   import { createEventDispatcher } from 'svelte'
   import calendar from '../plugin'
   import { saveUTC } from '../utils'
-  import CalendarSelector from './CalendarSelector.svelte'
-  import EventParticipants from './EventParticipants.svelte'
-  import EventReminders from './EventReminders.svelte'
-  import EventTimeEditor from './EventTimeEditor.svelte'
-  import EventTimeExtraButton from './EventTimeExtraButton.svelte'
-  import LocationEditor from './LocationEditor.svelte'
-  import ReccurancePopup from './ReccurancePopup.svelte'
-  import VisibilityEditor from './VisibilityEditor.svelte'
+  import EventManager from './EventManager.svelte'
 
   export let attachedTo: Ref<Doc> = calendar.ids.NoAttached
   export let attachedToClass: Ref<Class<Doc>> = calendar.class.Event
@@ -50,7 +31,6 @@
 
   const now = new Date()
   const defaultDuration = 60 * 60 * 1000
-  const allDayDuration = 24 * 60 * 60 * 1000 - 1
 
   let startDate =
     date === undefined ? now.getTime() : withTime ? date.getTime() : date.setHours(now.getHours(), now.getMinutes())
@@ -82,10 +62,6 @@
 
   const dispatch = createEventDispatcher()
   const client = getClient()
-
-  export function canClose (): boolean {
-    return title !== undefined && title.trim().length === 0 && participants.length === 0
-  }
 
   async function saveEvent () {
     let date: number | undefined
@@ -131,150 +107,23 @@
     }
     dispatch('close')
   }
-
-  async function allDayChangeHandler () {
-    if (allDay) {
-      startDate = new Date(startDate).setHours(0, 0, 0, 0)
-      if (dueDate - startDate < allDayDuration) dueDate = allDayDuration + startDate
-      else dueDate = new Date(dueDate).setHours(23, 59, 59, 999)
-    } else {
-      dueDate = startDate + defaultDuration
-    }
-  }
-
-  function setRecurrance () {
-    showPopup(ReccurancePopup, { rules, startDate }, undefined, (res) => {
-      if (res) {
-        rules = res
-      }
-    })
-  }
-
-  const manager = createFocusManager()
 </script>
 
-<FocusHandler {manager} />
-
-<div class="eventPopup-container">
-  <div class="header flex-between">
-    <EditBox
-      bind:value={title}
-      placeholder={calendar.string.EventTitlePlaceholder}
-      kind={'ghost-large'}
-      fullSize
-      focusable
-      autoFocus
-      focusIndex={10001}
-    />
-    <div class="flex-row-center gap-1 flex-no-shrink ml-3">
-      <Button id="card-more" focusIndex={10002} icon={IconMoreH} kind={'ghost'} size={'small'} on:click={() => {}} />
-      <Button
-        id="card-close"
-        focusIndex={10003}
-        icon={IconClose}
-        kind={'ghost'}
-        size={'small'}
-        on:click={() => {
-          dispatch('close')
-        }}
-      />
-    </div>
-  </div>
-  <Scroller thinScrollBars>
-    <div class="block first flex-no-shrink">
-      <EventTimeEditor {allDay} bind:startDate bind:dueDate {timeZone} focusIndex={10004} />
-      <EventTimeExtraButton
-        bind:allDay
-        bind:timeZone
-        bind:rules
-        on:repeat={setRecurrance}
-        on:allday={allDayChangeHandler}
-      />
-    </div>
-    <div class="block rightCropPadding">
-      <LocationEditor focusIndex={10010} bind:value={location} />
-      <EventParticipants focusIndex={10011} bind:participants bind:externalParticipants />
-    </div>
-    <div class="block row gap-1-5">
-      <div class="top-icon">
-        <Icon icon={calendar.icon.Description} size={'small'} />
-      </div>
-      <StyledTextBox
-        alwaysEdit={true}
-        isScrollable={false}
-        kind={'indented'}
-        showButtons={false}
-        focusIndex={10100}
-        placeholder={calendar.string.Description}
-        bind:content={description}
-      />
-    </div>
-    <div class="block rightCropPadding">
-      <CalendarSelector bind:value={space} focusIndex={10101} />
-      <div class="flex-row-center flex-gap-1">
-        <Icon icon={calendar.icon.Hidden} size={'small'} />
-        <VisibilityEditor bind:value={visibility} kind={'tertiary'} focusIndex={10102} withoutIcon />
-      </div>
-      <EventReminders bind:reminders focusIndex={10103} />
-    </div>
-  </Scroller>
-  <div class="antiDivider noMargin" />
-  <div class="flex-between p-5 flex-no-shrink">
-    <div />
-    <Button
-      kind="primary"
-      label={presentation.string.Create}
-      focusIndex={10104}
-      on:click={saveEvent}
-      disabled={title === ''}
-    />
-  </div>
-</div>
-
-<style lang="scss">
-  .eventPopup-container {
-    display: flex;
-    flex-direction: column;
-    max-width: 25rem;
-    min-width: 25rem;
-    min-height: 0;
-    background: var(--theme-popup-color);
-    border-radius: 1rem;
-    box-shadow: var(--theme-popup-shadow);
-
-    .header {
-      flex-shrink: 0;
-      padding: 0.75rem 0.75rem 0.5rem;
-    }
-
-    .block {
-      display: flex;
-      flex-shrink: 0;
-      min-width: 0;
-      min-height: 0;
-
-      &:not(:last-child) {
-        border-bottom: 1px solid var(--theme-divider-color);
-      }
-      &:not(.row) {
-        flex-direction: column;
-      }
-      &.first {
-        padding-top: 0;
-      }
-      &:not(.rightCropPadding) {
-        padding: 0.75rem 1.25rem;
-      }
-      &.rightCropPadding {
-        padding: 0.75rem 1rem 0.75rem 1.25rem;
-      }
-      &.row {
-        padding: 0 1.25rem 0.5rem;
-      }
-    }
-    .top-icon {
-      flex-shrink: 0;
-      margin-top: 1.375rem;
-    }
-  }
-</style>
+<EventManager
+  bind:title
+  bind:startDate
+  bind:dueDate
+  bind:visibility
+  bind:participants
+  bind:externalParticipants
+  bind:allDay
+  bind:reminders
+  bind:space
+  bind:timeZone
+  bind:description
+  bind:location
+  bind:rules
+  {duration}
+  on:save={saveEvent}
+  on:close={() => dispatch('close')}
+/>
