+++ plugins/calendar-resources/src/components/EditEvent.svelte
@@ -16,39 +16,18 @@
   import { Event, ReccuringEvent, ReccuringInstance, RecurringRule } from '@hcengineering/calendar'
   import { Person } from '@hcengineering/contact'
   import { DocumentUpdate, Ref } from '@hcengineering/core'
-  import presentation, { getClient } from '@hcengineering/presentation'
-  import { StyledTextBox } from '@hcengineering/text-editor'
-  import {
-    Button,
-    EditBox,
-    FocusHandler,
-    Icon,
-    IconClose,
-    createFocusManager,
-    getUserTimezone,
-    showPopup,
-    Scroller
-  } from '@hcengineering/ui'
+  import { getClient } from '@hcengineering/presentation'
+  import { getUserTimezone } from '@hcengineering/ui'
   import { deepEqual } from 'fast-equals'
   import { createEventDispatcher } from 'svelte'
   import calendar from '../plugin'
   import { isReadOnly, saveUTC, updateReccuringInstance } from '../utils'
-  import CalendarSelector from './CalendarSelector.svelte'
-  import EventParticipants from './EventParticipants.svelte'
-  import EventReminders from './EventReminders.svelte'
-  import EventTimeEditor from './EventTimeEditor.svelte'
-  import EventTimeExtraButton from './EventTimeExtraButton.svelte'
-  import LocationEditor from './LocationEditor.svelte'
-  import ReccurancePopup from './ReccurancePopup.svelte'
-  import VisibilityEditor from './VisibilityEditor.svelte'
+  import EventManager from './EventManager.svelte'
 
   export let object: Event
   $: readOnly = isReadOnly(object)
 
   let title = object.title
-
-  const defaultDuration = 60 * 60 * 1000
-  const allDayDuration = 24 * 60 * 60 * 1000 - 1
 
   let startDate = object.date
   const duration = object.dueDate - object.date
@@ -60,7 +39,7 @@
   let timeZone: string = object.timeZone ?? getUserTimezone()
 
   let description = object.description
-  let location = object.location
+  let location = object.location ?? ''
 
   let rules: RecurringRule[] = (object as ReccuringEvent).rules ?? []
 
@@ -69,10 +48,6 @@
 
   const dispatch = createEventDispatcher()
   const client = getClient()
-
-  export function canClose (): boolean {
-    return title !== undefined && title.trim().length === 0 && participants.length === 0
-  }
 
   async function saveEvent () {
     if (readOnly) {
@@ -131,157 +106,24 @@
     }
     dispatch('close')
   }
-
-  async function allDayChangeHandler () {
-    if (allDay) {
-      startDate = new Date(startDate).setHours(0, 0, 0, 0)
-      if (dueDate - startDate < allDayDuration) dueDate = allDayDuration + startDate
-      else dueDate = new Date(dueDate).setHours(23, 59, 59, 999)
-    } else {
-      dueDate = startDate + defaultDuration
-    }
-  }
-
-  function setRecurrance () {
-    if (readOnly) {
-      return
-    }
-    showPopup(ReccurancePopup, { rules, startDate }, undefined, (res) => {
-      if (res) {
-        rules = res
-      }
-    })
-  }
-
-  const manager = createFocusManager()
 </script>
 
-<FocusHandler {manager} />
-
-<div class="eventPopup-container">
-  <div class="header flex-between">
-    <EditBox
-      bind:value={title}
-      placeholder={calendar.string.EventTitlePlaceholder}
-      disabled={readOnly}
-      kind={'ghost-large'}
-      fullSize
-      focusable
-      autoFocus
-      focusIndex={10001}
-    />
-    <div class="flex-row-center gap-1 flex-no-shrink ml-3">
-      <Button
-        id="card-close"
-        focusIndex={10003}
-        icon={IconClose}
-        kind={'ghost'}
-        size={'small'}
-        on:click={() => {
-          dispatch('close')
-        }}
-      />
-    </div>
-  </div>
-  <Scroller thinScrollBars>
-    <div class="block first">
-      <EventTimeEditor {allDay} bind:startDate {timeZone} bind:dueDate disabled={readOnly} focusIndex={10004} />
-      <EventTimeExtraButton
-        bind:allDay
-        bind:timeZone
-        bind:rules
-        on:repeat={setRecurrance}
-        {readOnly}
-        on:allday={allDayChangeHandler}
-        noRepeat
-      />
-    </div>
-    <div class="block rightCropPadding">
-      <LocationEditor bind:value={location} focusIndex={10005} />
-      <EventParticipants bind:participants bind:externalParticipants disabled={readOnly} focusIndex={10006} />
-    </div>
-    <div class="block row gap-1-5">
-      <div class="top-icon">
-        <Icon icon={calendar.icon.Description} size={'small'} />
-      </div>
-      <StyledTextBox
-        alwaysEdit={true}
-        isScrollable={false}
-        focusIndex={10007}
-        kind={'indented'}
-        showButtons={false}
-        placeholder={calendar.string.Description}
-        bind:content={description}
-      />
-    </div>
-    <div class="divider" />
-    <div class="block rightCropPadding">
-      <CalendarSelector bind:value={space} focusIndex={10008} />
-      <div class="flex-row-center flex-gap-1">
-        <Icon icon={calendar.icon.Hidden} size={'small'} />
-        <VisibilityEditor bind:value={visibility} kind={'tertiary'} withoutIcon focusIndex={10009} />
-      </div>
-      <EventReminders bind:reminders focusIndex={10010} />
-    </div>
-  </Scroller>
-  <div class="antiDivider noMargin" />
-  <div class="flex-between p-5 flex-no-shrink">
-    <div />
-    <Button
-      kind="primary"
-      label={presentation.string.Save}
-      focusIndex={10011}
-      disabled={readOnly}
-      on:click={saveEvent}
-    />
-  </div>
-</div>
-
-<style lang="scss">
-  .eventPopup-container {
-    display: flex;
-    flex-direction: column;
-    max-width: 40rem;
-    min-width: 25rem;
-    min-height: 0;
-    background: var(--theme-popup-color);
-    border-radius: 1rem;
-    box-shadow: var(--theme-popup-shadow);
-
-    .header {
-      flex-shrink: 0;
-      padding: 0.75rem 0.75rem 0.5rem;
-    }
-
-    .block {
-      display: flex;
-      flex-shrink: 0;
-      padding: 0.75rem 1.25rem;
-      min-width: 0;
-      min-height: 0;
-
-      &:not(:last-child) {
-        border-bottom: 1px solid var(--theme-divider-color);
-      }
-      &:not(.row) {
-        flex-direction: column;
-      }
-      &.first {
-        padding-top: 0;
-      }
-      &:not(.rightCropPadding) {
-        padding: 0.75rem 1.25rem;
-      }
-      &.rightCropPadding {
-        padding: 0.75rem 1rem 0.75rem 1.25rem;
-      }
-      &.row {
-        padding: 0 1.25rem 0.5rem;
-      }
-    }
-    .top-icon {
-      flex-shrink: 0;
-      margin-top: 1.375rem;
-    }
-  }
-</style>
+<EventManager
+  bind:title
+  bind:startDate
+  bind:dueDate
+  bind:visibility
+  bind:participants
+  bind:externalParticipants
+  bind:allDay
+  bind:reminders
+  bind:space
+  bind:timeZone
+  bind:description
+  bind:location
+  bind:rules
+  {duration}
+  {readOnly}
+  on:save={saveEvent}
+  on:close={() => dispatch('close')}
+/>
